Move variance model out of mines microservice
