+++ python-backend/app/__init__.py
@@ -12,12 +12,12 @@ from flask_jwt_oidc.exceptions import AuthError
 from app.api.parties.namespace.parties import api as parties_api
 from app.api.applications.namespace.applications import api as applications_api
 from app.api.mines.namespace.mines import api as mines_api
-from app.api.mines.namespace.variances import api as variances_api
 from app.api.permits.namespace.permits import api as permits_api
 from app.api.documents.namespace.documents import api as document_api
 from app.api.document_manager.namespace.document_manager import api as document_manager_api
 from app.api.users.namespace.users import api as users_api
 from app.api.search.namespace.search import api as search_api
+from app.api.variances.namespace.variances import api as variances_api
 
 from app.commands import register_commands
 from app.config import Config
