+++ python-backend/app/api/mines/variances/resources/variance.py
@@ -7,7 +7,6 @@ from flask import request, current_app, Response
 from flask_restplus import Resource, fields
 from app.extensions import api, db
 
-from ..models.variance import Variance
 from ...mine.models.mine import Mine
 from ....documents.mines.models.mine_document import MineDocument
 from ....utils.access_decorators import (requires_any_of, MINE_VIEW, MINE_CREATE,
@@ -15,13 +14,15 @@ from ....utils.access_decorators import (requires_any_of, MINE_VIEW, MINE_CREATE
 from ....utils.resources_mixins import UserMixin, ErrorMixin
 from app.api.utils.custom_reqparser import CustomReqparser
 from app.api.mines.mine_api_models import VARIANCE_MODEL
+# TODO: Refactor to use API call
+from app.api.variances.models.variance import Variance
 # The need to access the guid -> id lookup forces an import as the id primary
 # key is not available via the API. The interal-only primary key +
 # cross-namespace foreign key constraints are interally inconsistent
 from app.api.users.core.models.core_user import CoreUser
 
 
-class VarianceResource(Resource, UserMixin, ErrorMixin):
+class MineVarianceResource(Resource, UserMixin, ErrorMixin):
     parser = CustomReqparser()
     parser.add_argument(
         'compliance_article_id',
