+++ python-backend/app/api/variances/models/variance.py
@@ -8,8 +8,8 @@ from sqlalchemy.orm import validates
 from app.extensions import db
 
 from .variance_application_status_code import VarianceApplicationStatusCode
-from ....utils.models_mixins import AuditMixin, Base
-from ....documents.variances.models.variance import VarianceDocumentXref
+from ...utils.models_mixins import AuditMixin, Base
+from ...documents.variances.models.variance import VarianceDocumentXref
 
 INVALID_GUID = 'Invalid guid.'
 INVALID_MINE_GUID = 'Invalid mine_guid.'
