+++ python-backend/app/api/variances/models/variance_application_status_code.py
@@ -1,7 +1,7 @@
 from sqlalchemy.schema import FetchedValue
 from app.extensions import db
 
-from ....utils.models_mixins import AuditMixin, Base
+from ...utils.models_mixins import AuditMixin, Base
 
 
 class VarianceApplicationStatusCode(AuditMixin, Base):
