+++ python-backend/tests/auth/test_expected_auth.py
@@ -18,10 +18,10 @@ from app.api.mines.mine.resources.mine_type_detail import MineTypeDetailResource
 from app.api.mines.mine.resources.mine_type import MineTypeResource, MineTypeListResource
 from app.api.mines.mine.resources.mine import MineResource, MineListSearch, MineListResource
 from app.api.mines.mine.resources.mine_map import MineMapResource
-from app.api.mines.variances.resources.variance import VarianceResource
-from app.api.mines.variances.resources.variance_list import VarianceListResource
-from app.api.mines.variances.resources.variance_document_upload import VarianceDocumentUploadResource
-from app.api.mines.variances.resources.variance_uploaded_documents import VarianceUploadedDocumentsResource
+from app.api.mines.variances.resources.variance import MineVarianceResource
+from app.api.mines.variances.resources.variance_list import MineVarianceListResource
+from app.api.mines.variances.resources.variance_document_upload import MineVarianceDocumentUploadResource
+from app.api.mines.variances.resources.variance_uploaded_documents import MineVarianceUploadedDocumentsResource
 from app.api.mines.region.resources.region import MineRegionResource
 from app.api.mines.status.resources.status import MineStatusResource
 from app.api.mines.tailings.resources.tailings import MineTailingsStorageFacilityListResource
@@ -55,21 +55,21 @@ from app.api.search.search.resources.simple_search import SimpleSearchResource
      (ExpectedDocumentUploadResource, "delete", [MINE_CREATE, MINESPACE_PROPONENT]),
      (ExpectedMineDocumentResource, "get", [MINE_VIEW]),
      (ExpectedMineDocumentResource, "post", [MINE_CREATE]),
-     (MineCommodityCodeResource, "get", [MINE_VIEW]), 
+     (MineCommodityCodeResource, "get", [MINE_VIEW]),
      (MineComplianceResource, "get", [MINE_VIEW]),
      (MineDisturbanceCodeResource, "get", [MINE_VIEW]),
      (MineDocumentResource, "get", [MINE_VIEW, MINESPACE_PROPONENT]),
-     (MineListResource, "get", [MINE_VIEW, MINESPACE_PROPONENT]), 
+     (MineListResource, "get", [MINE_VIEW, MINESPACE_PROPONENT]),
      (MineListResource, "post", [MINE_CREATE]),
      (MineListSearch, "get", [MINE_VIEW, MINESPACE_PROPONENT]),
      (MineMapResource, "get", [MINE_VIEW, MINESPACE_PROPONENT]),
      (MinePartyApptResource, "get", [MINE_VIEW]),
-     (MinePartyApptResource, "post", [MINE_CREATE]), 
+     (MinePartyApptResource, "post", [MINE_CREATE]),
      (MinePartyApptResource, "put", [MINE_CREATE]),
      (MinePartyApptResource, "delete", [MINE_CREATE]),
      (MinePartyApptTypeResource, "get", [MINE_VIEW]),
      (MineRegionResource, "get", [MINE_VIEW]),
-     (MineResource, "get", [MINE_VIEW, MINESPACE_PROPONENT]), 
+     (MineResource, "get", [MINE_VIEW, MINESPACE_PROPONENT]),
      (MineResource, "put", [MINE_CREATE]),
      (MineStatusResource, "get", [MINE_VIEW]),
      (MineTailingsStorageFacilityListResource, "get", [MINE_VIEW]),
@@ -79,6 +79,13 @@ from app.api.search.search.resources.simple_search import SimpleSearchResource
      (MineTypeDetailResource, "delete", [MINE_CREATE]),
      (MineTypeListResource, "post", [MINE_CREATE]),
      (MineTypeResource, "delete", [MINE_CREATE]),
+     (MineVarianceDocumentUploadResource, "post", [MINE_CREATE, MINESPACE_PROPONENT]),
+     (MineVarianceDocumentUploadResource, "put", [MINE_CREATE, MINESPACE_PROPONENT]),
+     (MineVarianceUploadedDocumentsResource, "delete", [MINE_CREATE, MINESPACE_PROPONENT]),
+     (MineVarianceListResource, "get", [MINE_VIEW, MINESPACE_PROPONENT]),
+     (MineVarianceListResource, "post", [MINE_CREATE, MINESPACE_PROPONENT]),
+     (MineVarianceResource, "get", [MINE_VIEW]),
+     (MineVarianceResource, "put", [MINE_CREATE, MINESPACE_PROPONENT]),
      (PartyListResource, "get", [MINE_VIEW, MINESPACE_PROPONENT]),
      (PartyListResource, "post", [MINE_CREATE]),
      (PartyResource, "get", [MINE_VIEW]),
@@ -95,18 +102,12 @@ from app.api.search.search.resources.simple_search import SimpleSearchResource
      (RequiredDocumentResource, "get", [MINE_VIEW]),
      (SearchResource, "get", [MINE_VIEW]),
      (SearchOptionsResource, "get", [MINE_VIEW]),
-     (SimpleSearchResource, "get", [MINE_VIEW]), 
+     (SimpleSearchResource, "get", [MINE_VIEW]),
      (MinespaceUserResource, 'get', [MINE_ADMIN]),
-     (MinespaceUserResource, 'post', [MINE_ADMIN]), 
+     (MinespaceUserResource, 'post', [MINE_ADMIN]),
      (MinespaceUserResource, 'delete', [MINE_ADMIN]),
      (MinespaceUserMineResource, 'post', [MINE_ADMIN]),
-     (MinespaceUserMineResource, 'delete', [MINE_ADMIN]),
-     (VarianceDocumentUploadResource, "post", [MINE_CREATE, MINESPACE_PROPONENT]),
-     (VarianceDocumentUploadResource, "put", [MINE_CREATE, MINESPACE_PROPONENT]),
-     (VarianceUploadedDocumentsResource, "delete", [MINE_CREATE, MINESPACE_PROPONENT]),
-     (VarianceListResource, "get", [MINE_VIEW, MINESPACE_PROPONENT]), (VarianceListResource, "post", [MINE_CREATE, MINESPACE_PROPONENT]),
-     (VarianceResource, "get", [MINE_VIEW]),
-     (VarianceResource, "put", [MINE_CREATE, MINESPACE_PROPONENT])])
+     (MinespaceUserMineResource, 'delete', [MINE_ADMIN])])
 def test_endpoint_auth(resource, method, expected_roles):
     endpoint = getattr(resource, method, None)
     assert endpoint != None, '{0} does not have a {1} method.'.format(resource, method.upper())
