+++ python-backend/tests/mines/variances/resources/test_variance_application_status_code_resource.py
@@ -3,7 +3,7 @@ import json
 from tests.factories import VarianceFactory, MineFactory
 from tests.status_code_gen import RandomComplianceArticleId
 from app.api.utils.custom_reqparser import DEFAULT_MISSING_REQUIRED
-from app.api.mines.variances.models.variance_application_status_code import VarianceApplicationStatusCode
+from app.api.variances.models.variance_application_status_code import VarianceApplicationStatusCode
 
 
 # GET
